ENH: add `never-cache` and `command` options
- `--never-cache`, `-n` option allows to specify patterns to match
files that will never be included in the cache. Even if the files were
in the cache already they will be ignored if the file matches the
provided pattern. Multiple patterns can be specified by specifying this
parameter multiple times.

- `command` option can be used to pass the cli command used to execute
persistify. If that command changes then the cache will be also be
ignored